Add catalog disable feature (#78)
Right now we require `__meta__.catalog.namespace` to be set to create
anything for an item. This feature allows `__meta__.catalog.disable` to
be set to specifically disable template and catalog item creation. This
is useful for items that are components of a combined catalog item that
should not be ordered separately.