Comment back in Pull Request when the run is done
This change, if applied, adds a feature in agnosticv-operator to report
back into all PRs, based on a certain kind of activity in commits.

It gets the list of PRs from the list of new commits using the convention setup in rhpds/agnosticv#7589 : All commits that start with the string `Auto-merge #...`.

We parse only the new commits, so we process the list of PRs only once and should not comment more than once for a change.

We will use that in conjunction with the integration branch.